requirements.txt
annotated-types==0.6.0
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
fastapi==0.110.1
gotrue==2.4.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.7
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
realtime==1.0.5
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0